requirements.txt
dj-database-url==0.5.0
Django==2.2.2
django-filter==2.1.0
pkg-resources==0.0.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
